+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ Change Log
 ----------
 
 
+0.7.4
+=====
+
+IDENTITY-ized Foursight; i.e. get secrets and other configuration data from the global application configuration (GAC)
+rather than having them encoded in the environment via the CloudFormation template (for the lambdas). C4-826.
+
+
 0.7.3
 =====